Calculation of combination index for binary mixtures

For single mixture data combination indices for effective doses as well as effects may be calculated and visualized.

CIcomp(mixProp, modelList, EDvec)

CIcompX(mixProp, modelList, EDvec, EDonly = FALSE)

plotFACI(effList, indAxis = c("ED", "EF"), caRef = TRUE,
showPoints = FALSE, add = FALSE, ylim, ...)

Arguments

mixProp
a numeric value between 0 and 1 specifying the mixture proportion/ratio for the single mixture considered.

modelList
a list contained 3 models fits using drm with the model fit for single mixture ratio being the first element, followed by the 2 model fits of the pure substances.

EDvec
a vector of numeric values between 0 and 100 (percentages) coresponding to the effect levels of interest.

EDonly
a logical value indicating whether or not only combination indices for effective doses should be calculated.

effList
a list returned by CIcompX.

indAxis
a character indicating whether effective doses ("ED") or effects ("EF") should be plotted.

caRef
a logical value indicating whether or not a reference line for concentration addition should be drawn.

showPoints
A logical value indicating whether or not estimated combination indices should be plotted.

add
a logical value specifying if the plot should be added to the existing plot.

ylim
a numeric vector of length 2 giving the range for the y axis.

...
additional graphical arguments.

Details

CIcomp calculates the classical combination index for effective doses whereas CIcompX calculates the combination index also for effects as proposed by Martin-Betancor et al. (2015); for details and examples using "drc" see the supplementary material of this paper. The function plotFACI may be used to visualize the calculated combination index as a function of the fraction affected.

Value

CIcomp returns a matrix which one row per ED value. Columns contain estimated combination indices, their standard errors and 95% confidence intervals, p-value for testing CI=1, estimated ED values for the mixture data and assuming concentration addition (CA) with corresponding standard errors.

CIcompX returns similar output both for effective doses and effects (as a list of matrices).

References

Martin-Betancor, K. and Ritz, C. and Fernandez-Pinas, F. and Leganes, F. and Rodea-Palomares, I. (2015) Defining an additivity framework for mixture research in inducible whole-cell biosensors, Scientific Reports 17200.

See also

Examples

## Fitting marginal models for the 2 pure substances
acidiq.0 <- drm(rgr ~ dose, data = subset(acidiq, pct == 999 | pct == 0), fct = LL.4())
acidiq.100 <- drm(rgr ~ dose, data = subset(acidiq, pct == 999 | pct == 100), fct = LL.4())

## Fitting model for single mixture with ratio 17:83
acidiq.17 <- drm(rgr ~ dose, data = subset(acidiq, pct == 17 | pct == 0), fct = LL.4())

## Calculation of combination indices based on ED10, ED20, ED50
CIcomp(0.17, list(acidiq.17, acidiq.0, acidiq.100), c(10, 20, 50))

#>      combInd         SE     lowCI   highCI    CAdiffp     ED.CA    SE.CA
#> 10 1.7180152 0.31407144 1.1024352 2.333595 0.02224534  76.91373 11.85583
#> 20 1.3421604 0.16702874 1.0147841 1.669537 0.04050985 140.38385 14.47436
#> 50 0.9035949 0.08440138 0.7381682 1.069022 0.25336168 382.44378 32.11935
#>      ED.mix   SE.mix
#> 10 132.1390 12.98677
#> 20 188.4176 13.13050
#> 50 345.5742 14.12771

## CI>1 significantly for ED10 and ED20, but not so for ED50