Potency of two herbicides

Data are from an experiment, comparing the potency of the two herbicides glyphosate and bentazone in white mustard Sinapis alba.

data(S.alba.comp)

Format

A data frame with 141 observations on the following 8 variables.

exp

a factor with levels ben1, ben2, gly1, gly2 indicating which experiment each observation belongs to.

herbicide

a factor with levels Bentazone Glyphosate (the two herbicides applied).

dose

a numeric vector containing the dose in g/ha.

drymatter

a numeric vector containing the response (dry matter in g/pot).

Tf

a numeric vector .

area

a numeric vector .

Fo

a numeric vector .

Fm

a numeric vector .

Details

The lower and upper limits for the two herbicides can be assumed identical, whereas slopes and ED50 values are different (in the log-logistic model).

Source

Christensen, M. G. and Teicher, H. B., and Streibig, J. C. (2003) Linking fluorescence induction curve and biomass in herbicide screening, Pest Management Science, 59, 1303–1310.

Examples

if (FALSE) { # \dontrun{
library(drc)

## Displaying the data
head(S.alba.comp)

## Fitting a four-parameter log-logistic model with common upper and lower limits
S.alba.comp.m1 <- drm(drymatter ~ dose, herbicide, data = S.alba.comp, fct = LL.4(),
pmodels = list(~herbicide, ~1, ~1, ~herbicide))
summary(S.alba.comp.m1)

## Plotting the fitted curves
plot(S.alba.comp.m1, xlab = "Dose (g/ha)", ylab = "Dry matter (g/pot)")
} # }