The two-parameter Weibull functions

'W1.2' is the two-parameter Weibull function where the lower limit is fixed at 0 and the upper limit is fixed at 1, mostly suitable for binomial/quantal responses.

W1.2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)

  W2.2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)

Arguments

upper
numeric value. The fixed, upper limit in the model. Default is 1.

fixed
numeric vector. Specifies which parameters are fixed and at what value they are fixed. NAs for parameter that are not fixed.

names
a vector of character strings giving the names of the parameters. The default is reasonable.

...
additional arguments to be passed from the convenience functions.

Details

The two-parameter Weibull type 1 model is given by the expression $$ f(x) = \exp(-\exp(b(\log(x)-\log(e)))).$$

The function is asymmetric about the inflection point, that is the parameter \(\exp(e)\).

Value

See weibull1.

Note

This function is for use with the function drm.

See also

Examples

## Fitting a two-parameter Weibull model
earthworms.m1 <- drm(number/total~dose, weights = total,
data = earthworms, fct = W1.2(), type = "binomial")

summary(earthworms.m1)

#> 
#> Model fitted: Weibull (type 1) with lower limit at 0 and upper limit at 1 (2 parms)
#> 
#> Parameter estimates:
#> 
#>               Estimate Std. Error t-value  p-value    
#> b:(Intercept) 0.475853   0.085228  5.5833 2.36e-08 ***
#> e:(Intercept) 0.179730   0.055978  3.2107 0.001324 ** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1