Effect of technical grade and commercially formulated auxin herbicides

MCPA, 2,4-D, mecorprop and dichorlprop were applied either as technical grades materials or as commercial formulations. Each experimental unit consisted of five 1-week old seedlings grown together in a pot of nutrient solution during 14 days.

data(auxins)

Format

A data frame with 150 observations on the following 5 variables.

dryweight

a numeric vector

dose

a numeric vector

replicate

a factor with 3 levels

herbicide

a factor with 5 levels

formulation

a factor with 2 levels

Details

Data are parts of a larger joint action experiment with various herbicides.

The eight herbicide preparations are naturally grouped into four pairs (herbicide:formulation) + control, and each pair of herbicides should have the same active ingredients but different formulation constituents, which were assumed to be biologically inert. The data consist of the 150 observations of dry weights, each observation being the weight of five plants grown in the same pot. All the eight herbicide preparations have essentially the same mode of action in the plant; they all act like the plant auxins, which are plant regulators that affect cell enlongation an other essential metabolic pathways. One of the objects of the experiment was to test if the response functions were identical except for a multiplicative factor in the dose. This is a necessary, but not a sufficient, condition for a similar mode of action for the herbicides.

Source

Streibig, J. C. (1987). Joint action of root-absorbed mixtures of auxin herbicides in Sinapis alba L. and barley (Hordeum vulgare L.) Weed Research, 27, 337–347.

References

Rudemo, M., Ruppert, D., and Streibig, J. C. (1989). Random-Effect Models in Nonlinear Regression with Applications to Bioassay. Biometrics, 45, 349–362.

Examples

if (FALSE) { # \dontrun{
library(drc)

## Displaying the data
head(auxins)

## Fitting a four-parameter log-logistic model with different curves per herbicide
auxins.m1 <- drm(dryweight ~ dose, herbicide, data = auxins, fct = LL.4())
summary(auxins.m1)

## Plotting the fitted curves
plot(auxins.m1, xlab = "Dose", ylab = "Dry weight")
} # }