Log-normal dose-response model

lnormal and the accompanying convenience functions provide a general framework for specifying the mean function of the decreasing or incresing log-normal dose-response model.

lnormal(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"),
  method = c("1", "2", "3", "4"), ssfct = NULL,
  fctName, fctText, loge = FALSE)

  LN.2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)

  LN.3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)

  LN.3u(upper = 1, fixed = c(NA, NA, NA), names = c("b", "c", "e"), ...)

  LN.4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)

Arguments

fixed
numeric vector. Specifies which parameters are fixed and at what value they are fixed. NAs for parameter that are not fixed.

names
vector of character strings giving the names of the parameters (should not contain ":"). The default is reasonable (see under 'Usage'). The order of the parameters is: b, c, d, e, f (see under 'Details' for the precise meaning of each parameter).

method
character string indicating the self starter function to use.

ssfct
a self starter function to be used.

fctName
character string used internally by convenience functions (optional).

fctText
character string used internally by convenience functions (optional).

loge
logical indicating whether or not ED50 or log(ED50) should be a parameter in the model. By default ED50 is a model parameter.

upper
numeric specifying the upper horizontal asymptote in the convenience function. The default is 1.

...
additional arguments to be passed from the convenience functions to lnormal.

Details

For the case where log(ED50), denoted \(e\) in the equation below, is a parameter in the model, the mean function is:

$$f(x) = c + (d-c)(\Phi(b(\log(x)-e)))$$

and the mean function is:

$$f(x) = c + (d-c)(\Phi(b(\log(x)-\log(e))))$$

in case ED50, which is also denoted \(e\), is a parameter in the model. If the former model is fitted any estimated ED values will need to be back-transformed subsequently in order to obtain effective doses on the original scale.

The mean functions above yield the same models as those described by Bruce and Versteeg (1992), but using a different parameterisation (among other things the natural logarithm is used).

For the case \(c=0\) and \(d=1\), the log-normal model reduces the classic probit model (Finney, 1971) with log dose as explanatory variable (mostly used for quantal data). This special case is available through the convenience function LN.2.

The case \(c=0\) is available as the function LN.3, whereas the LN.3u corresponds to the special case where the upper horizontal asymptote is fixed (default is 1). The full four-parameter model is available through LN.4.

Value

The value returned is a list containing the non-linear function, the self starter function and the parameter names.

References

Finney, D. J. (1971) Probit analysis, London: Cambridge University Press.

Bruce, R. D. and Versteeg, D. J. (1992) A statistical procedure for modeling continuous toxicity data, Environ. Toxicol. Chem., 11, 1485--1494.

Note

The function is for use with the function drm, but typically the convenience functions link{LN.2}, link{LN.3}, link{LN.3u}, and link{LN.4} should be used.

See also