3T3 mouse fibroblasts and NRU assay

The toxicity of sodium valproate was tested, using the 3T3 mouse fibroblasts and neutral red uptake (NRU) assay. 22 different experiments were performed independently in six laboratories, using eight concentration levels, each with six replicates on a 96-well plate. In addition, twelve measurements were taken for the solvent control.

data("mdra")

Format

A data frame with 1320 observations on the following 4 variables.

LabID

a factor with levels A B C D E F

ExperimentID

a factor with levels 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22

Concentration

a numeric vector

Response

a numeric vector

Source

http://biostatistics.dkfz.de/download/mdra/MDRA_ExampleData.csv

References

Clothier, R., Gomez-Lechon, M. J., Kinsner-Ovaskainen, A., Kopp-Schneider, A., O'Connor, J. E., Prieto, P., and Stanzel, S. (2013). Comparative analysis of eight cytotoxicity assays evaluated within the ACuteTox Project. Toxicology in vitro, 27(4):1347–1356.

Examples

if (FALSE) { # \dontrun{
data(mdra)

# fit a 3-parameter log-logistic mixed model
m <- medrm(Response ~ Concentration, data=mdra, fct=LL.3(), 
           random=d + e ~ 1|LabID/ExperimentID, 
           weights=varExp(form=~Concentration),
           start=c(2.13, 0.95, 0.02)) 
           
plot(m, logx=TRUE, ndose=250, ranef=TRUE) + 
     theme_classic()
} # }