The Cedergreen-Ritz-Streibig model

'CRS.4a', 'CRS.4b' and 'CRS.4c' provide the Cedergreen-Ritz-Streibig modified log-logistic model for describing hormesis with the lower limit equal to 0.

'UCRS.4a', 'UCRS.4b' and 'UCRS.4c' provide the Cedergreen-Ritz-Streibig modified log-logistic model for describing u-shaped hormesis with the lower limit equal to 0.

CRS.4a(names = c("b", "d", "e", "f"), ...)

  UCRS.4a(names = c("b", "d", "e", "f"), ...)

Arguments

names
a vector of character strings giving the names of the parameters. The default is reasonable (see above).

...
additional arguments to be passed from the convenience functions.

Details

The model is given by the expression $$ f(x) = 0 + \frac{d-0+f \exp(-1/x)}{1+\exp(b(\log(x)-\log(e)))}$$ which is a five-parameter model.

It is a modification of the four-parameter logistic curve to take hormesis into account.

The u-shaped model is given by the expression $$ f(x) = 0 + d - \frac{d-0+f \exp(-1/x^{\alpha})}{1+\exp(b(\log(x)-\log(e)))}$$

The a,b,c models are obtained by setting alpha equal to 1, 0.5 and 0.25, respectively.

Value

See cedergreen.

References

See the reference under cedergreen.

Note

This function is for use with the function drm.

See also

Examples

## Fitting modified logistic models
lettuce.crsm1 <- drm(lettuce[,c(2,1)], fct=CRS.4a())
summary(lettuce.crsm1)

#> 
#> Model fitted: Cedergreen-Ritz-Streibig with lower limit 0 (alpha=1) (4 parms)
#> 
#> Parameter estimates:
#> 
#>                Estimate Std. Error t-value   p-value    
#> b:(Intercept)  0.774519   0.248592  3.1156   0.01096 *  
#> d:(Intercept)  1.108705   0.078481 14.1270 6.212e-08 ***
#> e:(Intercept) 27.620019  30.307666  0.9113   0.38357    
#> f:(Intercept)  0.013090   0.417215  0.0314   0.97559    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error:
#> 
#>  0.1556406 (10 degrees of freedom)

ED(lettuce.crsm1, c(50))

#> 
#> Estimated effective doses
#> 
#>        Estimate Std. Error
#> e:1:50   28.436     11.618

## Need to explicitly specify that the upper limit
## is the reference in order to get ED10 and ED90 right
ED(lettuce.crsm1, c(10, 50, 90), reference = "upper")

#> 
#> Estimated effective doses
#> 
#>        Estimate Std. Error
#> e:1:10   1.7599     2.3788
#> e:1:50  28.4357    11.6181
#> e:1:90 479.2496   386.2069

lettuce.crsm2 <- drm(lettuce[,c(2,1)], fct=CRS.4b())
summary(lettuce.crsm2)

#> 
#> Model fitted: Cedergreen-Ritz-Streibig with lower limit 0 (alpha=.5) (4 parms)
#> 
#> Parameter estimates:
#> 
#>               Estimate Std. Error t-value   p-value    
#> b:(Intercept) 0.576206   0.090347  6.3777 8.060e-05 ***
#> d:(Intercept) 1.013816   0.097630 10.3842 1.124e-06 ***
#> e:(Intercept) 0.921496   2.937964  0.3137    0.7602    
#> f:(Intercept) 3.635793   6.503368  0.5591    0.5884    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error:
#> 
#>  0.1351142 (10 degrees of freedom)

ED(lettuce.crsm2, c(50))

#> 
#> Estimated effective doses
#> 
#>        Estimate Std. Error
#> e:1:50  26.3389     8.9005

lettuce.crsm3 <- drm(lettuce[,c(2,1)], fct=CRS.4c())
summary(lettuce.crsm3)

#> 
#> Model fitted: Cedergreen-Ritz-Streibig with lower limit 0 (alpha=.25) (4 parms)
#> 
#> Parameter estimates:
#> 
#>               Estimate Std. Error t-value   p-value    
#> b:(Intercept) 0.490601   0.137206  3.5757  0.005048 ** 
#> d:(Intercept) 0.973940   0.086907 11.2067 5.543e-07 ***
#> e:(Intercept) 1.378741   3.812958  0.3616  0.725179    
#> f:(Intercept) 2.935678   3.597993  0.8159  0.433550    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error:
#> 
#>  0.1235846 (10 degrees of freedom)

ED(lettuce.crsm3, c(50))

#> 
#> Estimated effective doses
#> 
#>        Estimate Std. Error
#> e:1:50   36.935     15.377