Cedergreen-Ritz-Streibig dose-reponse model for describing hormesis

'CRS.5a', 'CRS.5b' and 'CRS.5c' provide the Cedergreen-Ritz-Streibig modified log-logistic model for describing (inverse u-shaped or j-shaped) hormesis.

'UCRS.5a', 'UCRS.5b' and 'UCRS.5c' provide the Cedergreen-Ritz-Streibig modified log-logistic model for describing u-shaped hormesis.

CRS.5a(names = c("b", "c", "d", "e", "f"), ...)

  UCRS.5a(names = c("b", "c", "d", "e", "f"), ...)

Arguments

names
a vector of character strings giving the names of the parameters.

...
additional arguments to be passed from the convenience functions.

Details

The model function for inverse u-shaped hormetic patterns is

$$ f(x) = c + \frac{d-c+f \exp(-1/x^{\alpha})}{1+\exp(b(\log(x)-\log(e)))}$$,

which is a five-parameter model. It is a modification of the four-parameter log-logistic curve to take hormesis into account.

The parameters have the following interpretations

\(b\): Not direct interpretation

\(c\): Lower horizontal asymptote

\(d\): Upper horizontal asymptote

\(e\): Not direct interpretation

\(f\): Size of the hormesis effect: the larger the value the larger is the hormesis effect. \(f=0\) corresponds to no hormesis effect and the resulting model is the four-parameter log-logistic model. This parameter should be positive in order for the model to make sense.

The model function for u-shaped hormetic patterns is $$ f(x) = c + d - \frac{d-c+f \exp(-1/x^{\alpha})}{1+\exp(b(\log(x)-\log(e)))}$$

This model also simplifies to the four-parameter log-logistic model in case \(f=0\) (in a slightly different parameterization as compared to the one used in LL.4).

The models denoted a,b,c are obtained by fixing the alpha parameter at 1, 0.5 and 0.25, respectively.

Value

See cedergreen.

References

See the reference under cedergreen.

Note

This function is for use with the function drm.

See also

Examples

## Modified logistic model
lettuce.m1 <- drm(lettuce[,c(2,1)], fct=CRS.5a())
summary(lettuce.m1)

#> 
#> Model fitted: Cedergreen-Ritz-Streibig (alpha=1) (5 parms)
#> 
#> Parameter estimates:
#> 
#>               Estimate Std. Error t-value   p-value    
#> b:(Intercept) 1.334173   0.358675  3.7197  0.004773 ** 
#> c:(Intercept) 0.447962   0.080700  5.5510  0.000356 ***
#> d:(Intercept) 1.035658   0.077323 13.3940 3.004e-07 ***
#> e:(Intercept) 1.337869   1.185153  1.1289  0.288148    
#> f:(Intercept) 1.993259   2.017541  0.9880  0.348985    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error:
#> 
#>  0.1305067 (9 degrees of freedom)

ED(lettuce.m1, c(50))

#> 
#> Estimated effective doses
#> 
#>        Estimate Std. Error
#> e:1:50   5.5439     1.9480

lettuce.m2 <- drm(lettuce[,c(2,1)], fct=CRS.5b())
summary(lettuce.m2)

#> 
#> Model fitted: Cedergreen-Ritz-Streibig (alpha=.5) (5 parms)
#> 
#> Parameter estimates:
#> 
#>               Estimate Std. Error t-value   p-value    
#> b:(Intercept) 0.814273   0.414051  1.9666   0.08078 .  
#> c:(Intercept) 0.320033   0.165120  1.9382   0.08456 .  
#> d:(Intercept) 0.971182   0.081719 11.8844 8.361e-07 ***
#> e:(Intercept) 0.851872   2.178950  0.3910   0.70492    
#> f:(Intercept) 3.183596   5.641813  0.5643   0.58634    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error:
#> 
#>  0.1168455 (9 degrees of freedom)

ED(lettuce.m2, c(50))

#> 
#> Estimated effective doses
#> 
#>        Estimate Std. Error
#> e:1:50  11.4113     7.1885

lettuce.m3 <- drm(lettuce[,c(2,1)], fct=CRS.5c())
summary(lettuce.m3)

#> 
#> Model fitted: Cedergreen-Ritz-Streibig (alpha=.25) (5 parms)
#> 
#> Parameter estimates:
#> 
#>               Estimate Std. Error t-value   p-value    
#> b:(Intercept) 0.981945   0.559334  1.7556   0.11305    
#> c:(Intercept) 0.336670   0.182883  1.8409   0.09877 .  
#> d:(Intercept) 0.969845   0.088261 10.9883 1.624e-06 ***
#> e:(Intercept) 3.883893   2.462313  1.5773   0.14917    
#> f:(Intercept) 1.027934   0.766823  1.3405   0.21293    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error:
#> 
#>  0.1256841 (9 degrees of freedom)

ED(lettuce.m3, c(50))

#> 
#> Estimated effective doses
#> 
#>        Estimate Std. Error
#> e:1:50  11.4243     8.7214