The five-parameter log-logistic function

'LL.5' and 'LL2.5' provide the five-parameter log-logistic function, self starter function and names of the parameters.

LL.5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)

  l5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)

  LL2.5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)

Arguments

fixed
numeric vector. Specifies which parameters are fixed and at what value they are fixed. NAs for parameter that are not fixed.

names
a vector of character strings giving the names of the parameters. The default is reasonable.

...
Additional arguments (see llogistic).

Details

The five-parameter logistic function is given by the expression $$ f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-\log(e))))^f}$$

or in another parameterisation $$ f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-e)))^f}$$

The function is asymmetric for \(f\) different from 1.

Value

See llogistic.

References

Finney, D. J. (1979) Bioassay and the Practise of Statistical Inference, Int. Statist. Rev., 47, 1--12.

Note

This function is for use with the function drm.

See also

Examples

ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.5())
summary(ryegrass.m1)

#> 
#> Model fitted: Generalized log-logistic (ED50 as parameter) (5 parms)
#> 
#> Parameter estimates:
#> 
#>               Estimate Std. Error t-value p-value    
#> b:(Intercept)  3.92870    2.18882  1.7949 0.08860 .  
#> c:(Intercept)  0.32165    0.37943  0.8477 0.40715    
#> d:(Intercept)  7.76054    0.19319 40.1706 < 2e-16 ***
#> e:(Intercept)  2.21539    0.91883  2.4111 0.02619 *  
#> f:(Intercept)  0.46831    0.53020  0.8833 0.38813    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error:
#> 
#>  0.5270098 (19 degrees of freedom)