Maximum mean response

MAX estimates the maximum mean response and the dose at which it occurs.

MAX(object, lower = 1e-3, upper = 1000, pool = TRUE)

Arguments

object
an object of class 'drc'.

lower
numeric. Lower limit for bisection method. Need to be smaller than EDx level to be calculated.

upper
numeric. Upper limit for bisection method. Need to be larger than EDx level to be calculated.

pool
logical. If TRUE curves are pooled. Otherwise they are not. This argument only works for models with independently fitted curves as specified in drm.

Details

This function is only implemented for the built-in functions of class braincousens and cedergreen.

This function was used for obtaining the results on hormesis effect size reported in Cedergreen et al. (2005).

Value

A matrix with one row per curve in the data set and two columns: one containing the dose at which the maximum occurs and one containing the corresponding maximum response.

References

Cedergreen, N. and Ritz, C. and Streibig, J. C. (2005) Improved empirical models describing hormesis, Environmental Toxicology and Chemistry 24, 3166--3172.

Examples

## Fitting a Cedergreen-Ritz-Streibig model
lettuce.m1 <- drm(weight~conc, data = lettuce, fct = CRS.4c())

## Finding maximum average response and the corrresponding dose
MAX(lettuce.m1)

#>      Dose Response
#> 1 0.25587    1.178