Expected or predicted response

The function returns the expected or predicted response for specified dose values.

PR(object, xVec, ...)

Arguments

object
object of class drc obtaining fitting a dose-response model.

xVec
numeric vector of dose values.

…
additional arguments to be supplied to predict.drc. No effect at the moment.

Details

This function is a convenience function for easy access to predicted values.

Value

A numeric vector of predicted values or possibly a matrix of predicted values and corresponding standard errors.

See also

Examples

ryegrass.m1 <- drm(ryegrass, fct = LL.4())
PR(ryegrass.m1, c(5, 10))

#>         5        10 
#> 1.8523337 0.6888809 

ryegrass.m2 <- drm(ryegrass, fct = LL2.4())
PR(ryegrass.m2, c(5, 10))

#>         5        10 
#> 1.8523257 0.6888757 

spinach.m1 <- drm(SLOPE~DOSE, CURVE, data=spinach, fct = LL.4())
PR(spinach.m1, c(5, 10))

#>      Prediction         SE
#> 1:5   0.6849759 0.02733629
#> 1:10  0.5344815 0.02815873
#> 2:5   0.6849759 0.02733629
#> 2:10  0.5344815 0.02815873
#> 3:5   0.6849759 0.02733629
#> 3:10  0.5344815 0.02815873
#> 4:5   0.6849759 0.02733629
#> 4:10  0.5344815 0.02815873
#> 5:5   0.6849759 0.02733629
#> 5:10  0.5344815 0.02815873