The modified baro5 function

'baro5' allows specification of the baroreflex 5-parameter dose response function, under various constraints on the parameters.

baro5(fixed = c(NA, NA, NA, NA, NA), names = c("b1", "b2", "c", "d", "e"),
  method = c("1", "2", "3", "4"), ssfct = NULL)

Arguments

fixed
numeric vector. Specifies which parameters are fixed and at what value they are fixed. NAs for parameter that are not fixed.

names
a vector of character strings giving the names of the parameters (should not contain ":"). The order of the parameters is: b1, b2, c, d, e (see under 'Details').

method
character string indicating the self starter function to use.

ssfct
a self starter function to be used.

Details

The five-parameter function given by the expression

$$ y = c + \frac{d-c}{1+f\exp(b1(\log(x)-\log(e))) + (1-f)\exp(b2(\log(x)-\log(e)))}$$

$$ f = 1/( 1 + \exp((2b1b2/|b1+b2|)(\log(x)-\log(e))))$$

If the difference between the parameters b1 and b2 is different from 0 then the function is asymmetric.

Value

The value returned is a list containing the nonlinear model function, the self starter function and the parameter names.

References

Ricketts, J. H. and Head, G. A. (1999) A five-parameter logistic equation for investigating asymmetry of curvature in baroreflex studies. Am. J. Physiol. (Regulatory Integrative Comp. Physiol. 46), 277, 441--454.