Confidence Intervals for model parameters

Computes confidence intervals for one or more parameters in a model of class 'drc'.

# S3 method for drc
confint(object, parm, level = 0.95, pool = TRUE, ...)

Arguments

object
a model object of class 'drc'.

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
the confidence level required.

pool
logical. If TRUE curves are pooled. Otherwise they are not. This argument only works for models with independently fitted curves as specified in drm.

…
additional argument(s) for methods. Not used.

Details

For binomial and Poisson data the confidence intervals are based on the normal distribution, whereas t distributions are used of for continuous/quantitative data.

Value

A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in

Examples

## Fitting a four-parameter log-logistic model
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.4())

## Confidence intervals for all parameters
confint(ryegrass.m1)

#>                    2.5 %    97.5 %
#> b:(Intercept) 2.01211606 3.9523221
#> c:(Intercept) 0.03878752 0.9240389
#> d:(Intercept) 7.39961398 8.1863026
#> e:(Intercept) 2.67052621 3.4453837

## Confidence interval for a single parameter
confint(ryegrass.m1, "e")

#>                  2.5 %   97.5 %
#> e:(Intercept) 2.670526 3.445384