Model diagnostics for nonlinear dose-response models

Hat values (leverage values) and Cook's distance are provided for nonlinear dose-response model fits using the same formulas as in linear regression but based on the corresponding but approximate quantities available for nonlinear models.

# S3 method for drc
cooks.distance(model, ...)

  # S3 method for drc
hatvalues(model, ...)

Arguments

model
an object of class 'drc'.

…
additional arguments (not used).

Details

Hat values and Cook's distance are calculated using the formula given by Cook et al. (1986) and McCullagh and Nelder (1989).

The output values can be assessed in the same way as in linear regression.

Value

A vector of leverage values (hat values) or values of Cook's distance (one value per observation).

References

Cook, R. D. and Tsai, C.-L. and Wei, B. C. (1986) Bias in Nonlinear Regression, Biometrika 73, 615--623.

McCullagh, P. and Nelder, J. A. (1989) emphGeneralized Linear Models, Second edition, Chapman \& Hall/CRC.

Examples

ryegrass.LL.4 <- drm(rootl ~ conc, data = ryegrass, fct = LL.4())

hatvalues(ryegrass.LL.4)

#>          1          2          3          4          5          6          7 
#> 0.13332342 0.13332342 0.13332342 0.13332342 0.13332342 0.13332342 0.09539668 
#>          8          9         10         11         12         13         14 
#> 0.09539668 0.09539668 0.27948291 0.27948291 0.27948291 0.28191831 0.28191831 
#>         15         16         17         18         19         20         21 
#> 0.28191831 0.12467564 0.12467564 0.12467564 0.12949336 0.12949336 0.12949336 
#>         22         23         24 
#> 0.15571960 0.15571960 0.15571960 

cooks.distance(ryegrass.LL.4)

#>            1            2            3            4            5            6 
#> 7.453159e-03 7.044772e-03 4.714696e-02 4.844894e-02 2.870894e-02 4.723940e-03 
#>            7            8            9           10           11           12 
#> 6.453374e-02 4.817127e-02 3.034449e-03 1.086166e-01 1.026316e-03 1.159960e-01 
#>           13           14           15           16           17           18 
#> 6.500257e-01 1.505664e-02 6.990776e-01 8.318727e-03 1.370597e-03 1.649069e-03 
#>           19           20           21           22           23           24 
#> 3.231490e-03 6.070437e-05 1.244105e-02 1.159916e-02 1.468742e-02 4.949825e-04