The log-logistic function

'llogistic' provides a very general way of specifying log-logistic models, under various constraints on the parameters.

llogistic(fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("1", "2", "3", "4"), ssfct = NULL,
  fctName, fctText)

  llogistic2(fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  ss = c("1", "2", "3"), ssfct = NULL,
  fctName, fctText)

Arguments

fixed
numeric vector. Specifies which parameters are fixed and at what value they are fixed. NAs for parameter that are not fixed.

names
a vector of character strings giving the names of the parameters (should not contain ":"). The default is reasonable (see under 'Usage'). The order of the parameters is: b, c, d, e, f (see under 'Details').

method
character string indicating the self starter function to use.

ss
character string indicating the self starter function to use.

ssfct
a self starter function to be used.

fctName
optional character string used internally by convenience functions.

fctText
optional character string used internally by convenience functions.

Details

The default arguments yields the five-parameter log-logistic function given by the expression

$$ f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-\log(e))))^f}$$

If the parameter \(f\) differs from 1 then the function is asymmetric; otherwise it is symmetric (on log scale). This function is fitted using llogistic.

The log-logistic function with log(e) rather than e as a parameter, that is using the parameterisation

$$ f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-e)))^f}$$

is fitted using llogistic2.

Sometimes the log-logistic models are also called Hill models.

Value

The value returned is a list containing the nonlinear function, the self starter function and the parameter names.

References

Finney, D. J. (1979) Bioassay and the Practise of Statistical Inference, Int. Statist. Rev., 47, 1--12.

Seber, G. A. F. and Wild, C. J. (1989) Nonlinear Regression, New York: Wiley \& Sons (p. 330).

Note

The functions are for use with the function drm.

See also