Estimation of ED values using model-averaging

Estimates and confidence intervals for ED values are estimated using model-averaging.

maED(object, fctList = NULL, respLev, interval = c("none", "buckland", "kang"),
linreg = FALSE, clevel = NULL, level = 0.95, type = c("relative", "absolute"),
display = TRUE, na.rm = FALSE, extended = FALSE)

Arguments

object
an object of class 'drc'.

fctList
a list of non-linear functions to be compared.

respLev
a numeric vector containing the response levels.

interval
character string specifying the type of confidence intervals to be supplied. The default is "none". The choices "buckland" and "kang" are explained in the Details section.

linreg
logical indicating whether or not additionally a simple linear regression model should be fitted.

clevel
character string specifying the curve id in case on estimates for a specific curve or compound is requested. By default estimates are shown for all curves.

level
numeric. The level for the confidence intervals. The default is 0.95.

type
character string. Whether the specified response levels are absolute or relative (default).

display
logical. If TRUE results are displayed. Otherwise they are not (useful in simulations).

na.rm
logical indicating whether or not NA occurring during model fitting should be left out of subsequent calculations.

extended
logical specifying whether or not an extended output (including fit summaries) should be returned.

Details

Model-averaging of individual estimates is carried out as described by Buckland et al. (1997) and Kang et al. (2000) using AIC-based weights. The two approaches differ w.r.t. the calculation of confidence intervals: Buckland et al. (1997) provide an approximate variance formula under the assumption of perfectly correlated estimates (so, confidence intervals will tend to be too wide). Kang et al. (2000) use the model weights to calculate confidence limits as weighted means of the confidence limits for the individual fits; this procedure corresponds to using the standard error in Equation (3) given by Buckland et al. (1997) (assuming symmetric confidence intervals based on the same percentile).

Value

A matrix with two or more columns, containing the estimates and the corresponding estimated standard errors and possibly lower and upper confidence limits.

References

Buckland, S. T. and Burnham, K. P. and Augustin, N. H. (1997) Model Selection: An Integral Part of Inference, Biometrics 53, 603--618.

Kang, Seung-Ho and Kodell, Ralph L. and Chen, James J. (2000) Incorporating Model Uncertainties along with Data Uncertainties in Microbial Risk Assessment, Regulatory Toxicology and Pharmacology 32, 68--72.

See also

Examples

## Fitting an example dose-response model
ryegrass.m1 <- drm(rootl~conc, data = ryegrass, fct = LL.4())

## Comparing models (showing the AIC values)  
mselect(ryegrass.m1,
list(LL.5(), LN.4(), W1.4(), W2.4(), FPL.4(-1,1), FPL.4(-2,3), FPL.4(-0.5,0.5)))

#>                    logLik       IC Lack of fit   Res var
#> FPL.4(-0.5,0.5) -15.89038 41.78075           0 0.2641185
#> FPL.4(-1,1)     -15.90480 41.80959           0 0.2644360
#> W2.4            -15.91352 41.82703           0 0.2646283
#> LL.4            -16.15514 42.31029           0 0.2700107
#> LN.4            -16.29214 42.58429           0 0.2731110
#> FPL.4(-2,3)     -16.61493 43.22985           0 0.2805570
#> LL.5            -15.87828 43.75656           0 0.2777393
#> W1.4            -17.46720 44.93439           0 0.3012075

## Doing the actual model-averaging  
maED(ryegrass.m1,
list(LL.5(), LN.4(), W1.4(), W2.4(), FPL.4(-1,1), FPL.4(-2,3), FPL.4(-0.5,0.5)),
c(10, 50, 90))

#>                     ED10     ED50     ED90     Weight
#> LL.4            1.463706 3.057955 6.388640 0.14047308
#> LL.5            1.560325 3.023549 7.729713 0.06816147
#> LN.4            1.489188 3.044673 6.224889 0.12248817
#> W1.4            1.405979 3.088964 5.101022 0.03782468
#> W2.4            1.628278 2.996913 7.805803 0.17886712
#> FPL.4(-1,1)     1.540346 3.038790 7.086271 0.18043370
#> FPL.4(-2,3)     1.507055 3.063612 5.836831 0.08869758
#> FPL.4(-0.5,0.5) 1.531613 3.047967 7.204860 0.18305421
#> 

#>        Estimate
#> e:1:10 1.530770
#> e:1:50 3.039453
#> e:1:90 6.891117

## With confidence intervals according to Buckland et al. (1997)  
maED(ryegrass.m1,
list(LL.5(), LN.4(), W1.4(), W2.4(), FPL.4(-1,1), FPL.4(-2,3), FPL.4(-0.5,0.5)),
c(10, 50, 90), "buckland")

#>                     ED10     ED50     ED90     Weight
#> LL.4            1.463706 3.057955 6.388640 0.14047308
#> LL.5            1.560325 3.023549 7.729713 0.06816147
#> LN.4            1.489188 3.044673 6.224889 0.12248817
#> W1.4            1.405979 3.088964 5.101022 0.03782468
#> W2.4            1.628278 2.996913 7.805803 0.17886712
#> FPL.4(-1,1)     1.540346 3.038790 7.086271 0.18043370
#> FPL.4(-2,3)     1.507055 3.063612 5.836831 0.08869758
#> FPL.4(-0.5,0.5) 1.531613 3.047967 7.204860 0.18305421
#> 

#>        Estimate Std. Error    Lower    Upper
#> e:1:10 1.530770  0.1817370 1.174572 1.886968
#> e:1:50 3.039453  0.1922763 2.662598 3.416307
#> e:1:90 6.891117  1.4191777 4.109579 9.672654

## With confidence intervals according to Kang et al. (2000)
maED(ryegrass.m1,
list(LL.5(), LN.4(), W1.4(), W2.4(), FPL.4(-1,1), FPL.4(-2,3), FPL.4(-0.5,0.5)),
c(10, 50, 90), "kang")

#>                     ED10     ED50     ED90     Weight
#> LL.4            1.463706 3.057955 6.388640 0.14047308
#> LL.5            1.560325 3.023549 7.729713 0.06816147
#> LN.4            1.489188 3.044673 6.224889 0.12248817
#> W1.4            1.405979 3.088964 5.101022 0.03782468
#> W2.4            1.628278 2.996913 7.805803 0.17886712
#> FPL.4(-1,1)     1.540346 3.038790 7.086271 0.18043370
#> FPL.4(-2,3)     1.507055 3.063612 5.836831 0.08869758
#> FPL.4(-0.5,0.5) 1.531613 3.047967 7.204860 0.18305421
#> 

#>        Estimate    Lower    Upper
#> e:1:10 1.530770 1.172385 1.889155
#> e:1:50 3.039453 2.641383 3.437523
#> e:1:90 6.891117 4.377099 9.405134

## Comparing to model-averaged ED values with simple linear regression included
maED(ryegrass.m1,
list(LL.5(), LN.4(), W1.4(), W2.4(), FPL.4(-1,1), FPL.4(-2,3), FPL.4(-0.5,0.5)),
c(10, 50, 90), interval = "buckland", linreg = TRUE)

#>                     ED10      ED50      ED90       Weight
#> LL.4            1.463706  3.057955  6.388640 1.404731e-01
#> LL.5            1.560325  3.023549  7.729713 6.816147e-02
#> LN.4            1.489188  3.044673  6.224889 1.224882e-01
#> W1.4            1.405979  3.088964  5.101022 3.782468e-02
#> W2.4            1.628278  2.996913  7.805803 1.788671e-01
#> FPL.4(-1,1)     1.540346  3.038790  7.086271 1.804337e-01
#> FPL.4(-2,3)     1.507055  3.063612  5.836831 8.869758e-02
#> FPL.4(-0.5,0.5) 1.531613  3.047967  7.204860 1.830542e-01
#> Lin             2.407225 12.036124 21.665022 1.291283e-15
#> 

#>        Estimate Std. Error    Lower     Upper
#> e:1:10 1.530770  0.1969707 1.144714  1.916826
#> e:1:50 3.039453  1.0227988 1.034804  5.044102
#> e:1:90 6.891117  2.0331194 2.906276 10.875957

## Example with a model fit involving two compounds/curves
S.alba.m1 <- drm(DryMatter~Dose, Herbicide, data=S.alba, fct = LL.4(),
pmodels=data.frame(Herbicide,1,1,Herbicide))

## Model-averaged ED50 for both compounds
maED(S.alba.m1, list(LL.3(), LN.4()), 50)

#>          ED50       Weight
#> LL.4 66.89054 5.968947e-01
#> LL.3 60.10951 1.310984e-15
#> LN.4 67.18546 4.031053e-01
#> 
#>          ED50       Weight
#> LL.4 28.63235 5.968947e-01
#> LL.3 60.10951 1.310984e-15
#> LN.4 28.64257 4.031053e-01
#> 

#>                 Estimate
#> e:Glyphosate:50 67.00943
#> e:Bentazone:50  28.63647

## Model-averaged ED50 only for one compound (glyphosate)
maED(S.alba.m1, list(LL.3(), LN.4()), 50, clevel="Glyphosate")

#>          ED50       Weight
#> LL.4 66.89054 5.968947e-01
#> LL.3 60.10951 1.310984e-15
#> LN.4 67.18546 4.031053e-01
#> 

#>                 Estimate
#> e:Glyphosate:50 67.00943

## With confidence intervals
maED(S.alba.m1, list(LL.3(), LN.4()), 50, interval="buckland")

#>          ED50       Weight
#> LL.4 66.89054 5.968947e-01
#> LL.3 60.10951 1.310984e-15
#> LN.4 67.18546 4.031053e-01
#> 
#>          ED50       Weight
#> LL.4 28.63235 5.968947e-01
#> LL.3 60.10951 1.310984e-15
#> LN.4 28.64257 4.031053e-01
#> 

#>                 Estimate Std. Error     Lower    Upper
#> e:Glyphosate:50 67.00943   6.439439 54.388360 79.63050
#> e:Bentazone:50  28.63647  10.677172  7.709599 49.56334

## For comparison model-specific confidence intervals
ED(S.alba.m1, 50, interval="delta")  # wider!

#> 
#> Estimated effective doses
#> 
#>                 Estimate Std. Error   Lower   Upper
#> e:Bentazone:50   28.6324     2.0381 24.5583 32.7065
#> e:Glyphosate:50  66.8905     5.9688 54.9590 78.8220